pyproject.toml
[project]
name = "github-mcp"
version = "0.1.0"
description = "GitHub MCP server for Claude"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "pygithub>=2.1.1",
    "python-dotenv>=1.0.0"
]

[project.scripts]
github-mcp = "github_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

